> Cambio tipo nel metodo calcola

+++ OOP-Poject-TwitterTimeline/src/main/java/com/exam/Abstract/AbstractStatistic.java
@@ -13,6 +13,6 @@ public abstract class AbstractStatistic {
 		}
 
 
-		public abstract int calcola(ArrayList<Tweet> vett, String field);
+		public abstract double calcola(ArrayList<Tweet> vett, String field);
 
 }
